+++ src/freenet/client/async/USKManager.java
@@ -126,7 +126,7 @@ public class USKManager {
 	public synchronized long lookupKnownGood(USK usk) {
 		Long l = latestKnownGoodByClearUSK.get(usk.clearCopy());
 		if(l != null)
-			return l.longValue();
+			return l;
 		else return -1;
 	}
 
@@ -138,7 +138,7 @@ public class USKManager {
 	public synchronized long lookupLatestSlot(USK usk) {
 		Long l = latestSlotByClearUSK.get(usk.clearCopy());
 		if(l != null)
-			return l.longValue();
+			return l;
 		else return -1;
 	}
 
@@ -431,7 +431,7 @@ public class USKManager {
 		synchronized(this) {
 			Long l = latestKnownGoodByClearUSK.get(clear);
 			if(logMINOR) Logger.minor(this, "Old known good: "+l);
-			if((l == null) || (number > l.longValue())) {
+			if((l == null) || (number > l)) {
 				l = number;
 				latestKnownGoodByClearUSK.put(clear, l);
 				if(logMINOR) Logger.minor(this, "Put "+number);
@@ -440,7 +440,7 @@ public class USKManager {
 			
 			l = latestSlotByClearUSK.get(clear);
 			if(logMINOR) Logger.minor(this, "Old slot: "+l);
-			if((l == null) || (number > l.longValue())) {
+			if((l == null) || (number > l)) {
 				l = number;
 				latestSlotByClearUSK.put(clear, l);
 				if(logMINOR) Logger.minor(this, "Put "+number);
@@ -471,7 +471,7 @@ public class USKManager {
 		synchronized(this) {
 			Long l = latestSlotByClearUSK.get(clear);
 			if(logMINOR) Logger.minor(this, "Old slot: "+l);
-			if((l == null) || (number > l.longValue())) {
+			if((l == null) || (number > l)) {
 				l = number;
 				latestSlotByClearUSK.put(clear, l);
 				if(logMINOR) Logger.minor(this, "Put "+number);
